Delete

+++ database/migrations/2024_04_11_120623_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->double('priceHt');
+            $table->float('priceHt');
             $table->datetime('creationDate');
             $table->datetime('dateUpdate')->nullable();
